Fix window corruption with Matrox Millennium when using 9-30-1998 Win9x driver
The Developer Specification for the MGA 1064SG (Matrox Millennium) states for the
SGNZERO bit of DWGCTRL that "Writing to the SGN register when sgnzero = 1 will
produce unpredictable results". However the 9-30-1998 Win9x driver will do exactly
this when programming FBITBLT commands. To accommodate this, handling of SGNZERO and
the other *ZERO bits have been moved to take effect immediately on writing to
DWGCTRL rather than when a blit actually starts.

This fixes corruption when moving windows in 32 bpp mode on the 9-30-1998 driver.